Removed a lot of nuget packages relationship. The Code is not rely on project reference for easily debugging.

+++ 00 - Tony/00FreemindBase/20_fxCollections/DictionarySlimP.cs
@@ -289,7 +289,7 @@ namespace fx.Collections
             Debug.Assert(_entries.Length == _count || _entries.Length == 1); // We only copy _count, so if it's longer we will miss some
             int count = _count;
             int newSize = _entries.Length * 2;
-            if ((uint)newSize > (uint)int.MaxValue) // uint cast handles overflow
+            if ((uint)newSize > int.MaxValue ) // uint cast handles overflow
                 throw new InvalidOperationException( "Arg_HTCapacityOverflow");
 
             int[] buckets;
